fix: move server-request logic out of index.ts
